+++ mlxtend/preprocessing/__init__.py
@@ -4,14 +4,14 @@
 #
 # License: BSD 3 clause
 
-from .transformer import TransformerObj
 from .mean_centering import MeanCenterer
 from .shuffle import shuffle_arrays_unison
 from .scaling import minmax_scaling
 from .scaling import standardize
 from .dense_transformer import DenseTransformer
+from .copy_transformer import CopyTransformer
 from .onehot import one_hot
 
 
-__all__ = ["MeanCenterer", "shuffle_arrays_unison", "TransformerObj",
+__all__ = ["MeanCenterer", "shuffle_arrays_unison", "CopyTransformer",
            "minmax_scaling", "standardize", "DenseTransformer", "one_hot"]